Add Autocompletion to Option Search
Add a python file to generate an option_autocomplete.json file which will
be used as the choices for the search toolbox.

Let the autocomplete trigger also set the target_option so that the path
highlighting and opening can be completed.

Change the type of value used to match the target options from the name to
the entire node.