Begin integration of VA Visualizations
Add simple links to the top header which point to the websites for the
"Business Information Model" and the "Hybrid Information Model"

Add links to each package pointing to a relevant page in the HIM
framework.  If no page is available, generate a link to the top level
page.  The links to generate are found in the himData.json file.