Fix non-package dependency display
Fix the parsing of the package informaiton such that packages which are not
in the same package as the root install are able to show the appropriate
dependencies.

Fix the install_information.json to be keyed to the install name instead
of the index of the array.  Requires an update to the VistADataParse
branch.

OSEHRA-Id: http://issues.osehra.org/browse/VIVIAN-159