Add panning to patch dependency
Change the style of the tree to not be constrained by size, but to ensure
that the spacing between the nodes is consistent.

Add the capability to drag the patch depdendency tree around to be able to
see all parts of the visualization.  Ensure that the default position of
each tree is maintained to allow the reset to bring it back to a single
position.