Add "All Patches" option to Patch Dependency
Add an "package" option which will allow the user to select from all
available patches when viewing the dependencies of the patches.

This will let someone who doesn't know the corresponding package to still
find information about the patch.

This patch requires a change to the install_information.json file which is
generated by the following Gerrit patch:
http://review.code.osehra.org/#/c/851/